#119 split runtime and test modules
