+++ src/errors.ts
@@ -1,0 +1,3 @@
+export function errorMessage(e: unknown): string {
+  return e instanceof Error ? e.message : String(e);
+}
