+++ tsconfig.json
@@ -4,9 +4,9 @@
     "module": "NodeNext",
     "moduleResolution": "NodeNext",
     "outDir": "dist",
-    "rootDir": ".",
+    "rootDir": "src",
     "strict": true,
     "types": ["node"]
   },
-  "include": ["index.ts"]
+  "include": ["src/**/*.ts"]
 }
